sl_compiler::compile

Module compile_let

Functions

add_right_side_exp 🔒

compile_let 🔒

compile_let_value 🔒

Compile a value “out of the way” of the let bindings and then set the target register. For calls that use multiple regs and destructuring lets in use registers can get walked over without this. Use SET not MOV, we don’t want to replace a value that has been captured and SET honors this, MOV does not.

compile_let_while 🔒

compile_right_exps 🔒

let_inner 🔒

let_while_inner 🔒

Type Aliases

RightSideExp 🔒