sl_compiler
Module load_eval

Constants

SLSHRC

Functions

add_load_builtins

apply_callable

Call lambda with args, this is re-entrant.

exec_unrooted_chunk

Execute a chunk that may not be rooted, will make sure any allocated consts don’t get GCed out from under it…

expand_tilde

Takes a PathBuf, if it contains ~ then replaces it with HOME and returns the new PathBuf, else returns path. If path is not utf-8 then it will not expand ~.

get_load_name

load

load_internal

load_one_expression

run_reader

With the given reader, for each sexp compile then load and execute.