sl_compiler

Module load_eval

Constants

BUILTINS

COLORS_LISP 🔒

COLORS_LISP_NAME

CORE_LISP 🔒

CORE_LISP_NAME

ITER_LISP 🔒

ITER_LISP_NAME

SLSHRC

SLSHRC_NAME

TEST_LISP 🔒

TEST_LISP_NAME

Functions

add_compiler_builtin 🔒

add_load_builtins

apply 🔒

apply_callable

Call lambda with args, this is re-entrant.

apply_inner 🔒

check_builtins 🔒

contains_list 🔒

eval 🔒

Builtin eval implementation. Tries to avoid compilation when possible (uses apply machinery).

eval_exp 🔒

exec_unrooted_chunk

Execute a chunk that may not be rooted, will make sure any allocated consts don’t get GCed out from under it…

expand_tilde

Takes a PathBuf, if it contains ~ then replaces it with HOME and returns the new PathBuf, else returns path. If path is not utf-8 then it will not expand ~.

find_first_instance_of_file_in_load_path 🔒

Find file name the first time it appears in the global variable load-path.

from_utf8 🔒

get_load_name

load

load_internal

load_one_expression

quote_list 🔒

read 🔒

read_all 🔒

reader_for_file 🔒

run_reader

With the given reader, for each sexp compile then load and execute.

run_script 🔒